ptrace: handle a task saved_sigmask for PTRACE_{GET,SET}SIGMASK
There are a few system calls (pselect, ppoll, etc) which replace a task
sigmask while they are running in a kernel.

When one of this syscalls is started, a task sigmask is saved in
taks->saved_sigmask and a syscall mask is set. ptrace traps a task
before restoring the saved_sigmask, and now PTRACE_GETSIGMASK returns the
syscall sigmask and PTRACE_SETSIGMASK does nothing, because its sigmask
is replaced by saved_sigmask.