[RFC] mm: allow temporary switching mm and run in its context
This change introduces a new system call:

long process_vm_exec(pid_t, pid, struct sigcontext uctx, unsigned long flags)

This syscall call changes the current address space on an address space
of a specified task and run the current process with registers provided
in uctx. It stops on entry to the next system call or a signal. Than it
saves a process state in uctx, switch back the address space and returns
from the system call.

On success, process_vm_exec returns 0 if the stop has been triggered by a
sytem call or it returns a signal number if the stop has been triggred
by a signal.
On error, it returns a error code.

The behaviour of this system call is similar with PTRACE_SYSEMU but
everything is happing in context of one process and so process_vm_exec
show a better performance.

PTRACE_SYSEMU is primarily used to implement sandboxes (application
kernels) like user-mode linux or gVisor. These type of sandboxes
intercepts applications system calls and acts as the guest kernel.

A simple benchmark, where a "tracee" process executes systems calls in a
loop, and a "tracer" process traps syscalls and handles them just
incrementing the tracee instruction pointer to skip the syscall
instruction, shows that process_vm_exec works six times faster than
PTRACE_SYSEMU.

Signed-off-by: Andrei Vagin <avagin@gmail.com>